# doc/api.yaml
openapi: 3.0.0
paths:
  /user:
    get:
      operationId: UserController_findAll
      summary: Get all users
      description: Gets all users
      parameters: []
      responses:
        '200':
          description: Users have been successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDto'
      tags: &ref_0
        - user
      security: &ref_1
        - bearer: []
    post:
      operationId: UserController_create
      summary: Create user
      description: Creates a new user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: The user has been successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Request body does not contain required fields
        '409':
          description: User with provided login already exists
      tags: *ref_0
      security: *ref_1
  /user/{id}:
    get:
      operationId: UserController_findOne
      summary: Get single user by id
      description: Get single user by id
      parameters:
        - name: id
          required: true
          in: path
          description: user id (uuid v4)
          example: e83cd69d-d1ef-4770-835e-5eb0537cc5c9
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: The user has been successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Provided user id is invalid (not uuid)
        '404':
          description: User with provided id was not found
      tags: *ref_0
      security: *ref_1
    put:
      operationId: UserController_updatePassword
      summary: Update user's password
      description: Updates a user's password by ID
      parameters:
        - name: id
          required: true
          in: path
          description: user id (uuid v4)
          example: e83cd69d-d1ef-4770-835e-5eb0537cc5c9
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
      responses:
        '200':
          description: User's password has been successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Provided user id is invalid (not uuid)
        '403':
          description: Provided user's password is wrong
        '404':
          description: User with provided id was not found
      tags: *ref_0
      security: *ref_1
    delete:
      operationId: UserController_remove
      summary: Delete user
      description: Deletes user by ID
      parameters:
        - name: id
          required: true
          in: path
          description: user id (uuid v4)
          example: e83cd69d-d1ef-4770-835e-5eb0537cc5c9
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: User has been successfully deleted
        '400':
          description: Provided user id is invalid (not uuid)
        '404':
          description: User with provided id was not found
      tags: *ref_0
      security: *ref_1
...
  /track:
    get:
      operationId: TrackController_findAll
      summary: Get tracks list
      description: Gets all library tracks list
      parameters: []
      responses:
        '200':
          description: Tracks have been successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrackResponseDto'
      tags: &ref_6
        - track
      security: &ref_7
        - bearer: []
    post:
      operationId: TrackController_create
      summary: Add new track
      description: Add new track information
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackDto'
      responses:
        '201':
          description: The track has been successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackResponseDto'
        '400':
          description: >-
            Request body does not contain required fields or one of ids in DTO
            is invalid or points to non-existing entity
      tags: *ref_6
      security: *ref_7
  /track/{id}:
    get:
      operationId: TrackController_findOne
      summary: Get single track by id
      description: Get single track by id
      parameters:
        - name: id
          required: true
          in: path
          description: track id (uuid v4)
          example: 11035eda-0858-43bc-9ab7-9d4ba0da4e09
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: The track has been successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackResponseDto'
        '400':
          description: Provided id of track is invalid (not uuid)
        '404':
          description: Track with provided id was not found
      tags: *ref_6
      security: *ref_7
    put:
      operationId: TrackController_updateInfo
      summary: Update track information
      description: Update library track information by UUID
      parameters:
        - name: id
          required: true
          in: path
          description: track id (uuid v4)
          example: 3a36a0e8-186e-4861-940b-12326e4f4691
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackInfoDto'
      responses:
        '200':
          description: Track's info has been successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackResponseDto'
        '400':
          description: >-
            Provided id of track is invalid (not uuid) or one of ids in DTO is
            invalid or points to non-existing entity
        '404':
          description: Track with provided id was not found
      tags: *ref_6
      security: *ref_7
    delete:
      operationId: TrackController_remove
      summary: Delete track
      description: Delete track from library
      parameters:
        - name: id
          required: true
          in: path
          description: track id (uuid v4)
          example: a7f382c5-a607-4cb4-be24-c2b669c1017d
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Track has been successfully deleted
        '400':
          description: Provided id of track is invalid (not uuid)
        '404':
          description: Track with provided id was not found
      tags: *ref_6
      security: *ref_7
...
  /favs/track/{id}:
    post:
      operationId: FavsController_addTrack
      summary: Add track to the favorites
      description: Add track to the favorites
      parameters:
        - name: id
          required: true
          in: path
          description: track id (uuid v4)
          example: ba414b60-d562-4c80-af08-028d43b39f0a
          schema:
            format: uuid
            type: string
      responses:
        '201':
          description: The track has been successfully added to favorites
        '400':
          description: Provided id of track is invalid (not uuid)
        '422':
          description: Track with provided id was not found
      tags: *ref_8
      security: *ref_9
    delete:
      operationId: FavsController_removeTrack
      summary: Delete track from favorites
      description: Delete track from favorites
      parameters:
        - name: id
          required: true
          in: path
          description: track id (uuid v4)
          example: accb9d24-8192-419e-8e06-01071f90c0e2
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Track has been successfully deleted from the favorites
        '400':
          description: Provided id of track is invalid (not uuid)
        '404':
          description: Track with provided id was not found in the favorites
      tags: *ref_8
      security: *ref_9
...
info:
  title: Home Library Service
  description: 'Home Library Service: RS School NodeJS 2023 Q2 - Weeks 6-9 Task'
  version: 0.0.1
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
          description: user ID (uuid v4)
          example: 83ef2c0c-967e-46bb-b748-6614bc99b7b4
        login:
          type: string
          description: user login
          example: johndoe
        version:
          type: number
          minimum: 1
          description: user version (integer number, increments on update)
          example: 1
        createdAt:
          type: number
          description: timestamp of user creation
          example: 1690366188813
        updatedAt:
          type: number
          description: timestamp of last user update
          example: 1690366221355
      required:
        - id
        - login
        - version
        - createdAt
        - updatedAt
    CreateUserDto:
      type: object
      properties:
        login:
          type: object
          default: ''
          description: user login
          example: johndoe
        password:
          type: object
          default: ''
          description: user password
          example: abracadabra
      required:
        - login
        - password
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: object
          default: ''
          description: old password of the user
          example: abracadabra
        newPassword:
          type: object
          default: ''
          description: new password of the user
          example: d@n*9QRe6KPE9uv7
      required:
        - oldPassword
        - newPassword
...
    TrackResponseDto:
      type: object
      properties:
        id:
          type: string
          description: track ID (uuid v4)
          example: c81e086c-2cd9-4a9f-a68e-fa6c92214d52
        name:
          type: string
          description: track name
          example: Ghost Identifier
        artistId:
          type: string
          nullable: true
          description: ID of the artist who created this track (or null)
          example: 82e5e3a7-5f9a-4de4-9bfd-1398d3c0571e
        albumId:
          type: string
          nullable: true
          description: ID of the album this track belongs to (or null)
          example: 48d83402-b63c-4d23-abdf-43094e0d05b9
        duration:
          type: number
          minimum: 1
          description: track duration in seconds
          example: 353
      required:
        - id
        - name
        - artistId
        - albumId
        - duration
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: track name
          example: An Imperfect System
        artistId:
          type: string
          nullable: true
          description: ID of the artist who created this track (or null)
          example: d7733d90-6119-4347-95d2-ed85678a3e42
        albumId:
          type: string
          nullable: true
          description: ID of the album this track belongs to (or null)
          example: bbf3d2ef-ed3c-4c5b-b156-9d9f2264c5a0
        duration:
          type: number
          minimum: 1
          description: track duration in seconds
          example: 484
      required:
        - name
        - artistId
        - albumId
        - duration
    UpdateTrackInfoDto:
      type: object
      properties:
        name:
          type: string
          description: track name
          example: Mont Da Anaon
        artistId:
          type: string
          nullable: true
          description: ID of the artist who created this track (or null)
          example: 39dd467c-f47c-4387-a181-f5323fdf907d
        albumId:
          type: string
          nullable: true
          description: ID of the album this track belongs to (or null)
          example: ed42b070-738c-4c75-8acb-a7d79d537b38
        duration:
          type: number
          minimum: 1
          description: track duration in seconds
          example: 350
      required:
        - name
        - artistId
        - albumId
        - duration
...